+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated the doc structure and minor language fixes
 - Adding AssetsCache.readBinaryFile
 - Splitting debugMode from recordFps mode
+- Fix animations example
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice